Change `buff_size` param, `unsigned` to `size_t`
Notes:
- Fixes #210

Refs:
- <https://github.com/emilk/loguru/issues/210>

+++ loguru.cpp
@@ -710,7 +710,7 @@ namespace loguru
 		#endif // _WIN32
 	}
 
-	void suggest_log_path(const char* prefix, char* buff, unsigned buff_size)
+	void suggest_log_path(const char* prefix, char* buff, size_t buff_size)
 	{
 		if (prefix[0] == '~') {
 			snprintf(buff, buff_size - 1, "%s%s", home_dir(), prefix + 1);
